centralized app testing framework
